chore(Icon): move the file for prop type check
